[TASK] Move the subclass test to unit tests
This improves test performance.

Fixes #416
